update(exec): sep get command func to get_exec_command.c. and add exec.h, update related files
