+++ Makefile
@@ -18,6 +18,7 @@ SRCS		+=	$(EXEC_DIR)/check_command.c \
 				$(EXEC_DIR)/child_process.c \
 				$(EXEC_DIR)/exec.c \
 				$(EXEC_DIR)/exec_builtin_in_parent_proc.c \
+				$(EXEC_DIR)/get_exec_command.c \
 				$(EXEC_DIR)/init.c \
 				$(EXEC_DIR)/parent_pipes.c \
 				$(EXEC_DIR)/parent_process.c \
