+++ srcs/builtin/call_builtin_func.c
@@ -1,6 +1,7 @@
 #include "ft_builtin.h"
-#include "minishell.h"
+#include "exec.h"
 #include "ft_string.h"
+#include "minishell.h"
 
 bool	is_single_builtin(t_deque_node *node, const char *cmd)
 {
