+++ srcs/main.c
@@ -1,5 +1,6 @@
 #include "minishell.h"
 #include "tokenize.h"
+#include "exec.h"
 #include "deque.h"
 #include "ft_dprintf.h"
 #include "ft_builtin.h"
